gcc: patch *.txt data files, with minimum impact on sources
gcc do not allow to store a structure containing a flexible array member in an
array.

Also, sizeof (s), with s a structure containing a flexible array member will
return the offset of the flexible array member, not the actual size.

This patch handle those problems with minimum impact on sources.