gcc: automatically compute user flash size
This is done at link time.  As a consequence, the user flash size is not
available at compile time, which is a problem for the FILEHEADER type.  This
is solved by using a maximum size so that the static Header variable is big
enough.

It would not have been a problem if this variable was put on the stack.